+++ Tests/BugsnagEventTests.m
@@ -118,9 +118,9 @@
     XCTAssertEqualObjects(@"Exception",
                           payload[@"exceptions"][0][@"errorClass"]);
     XCTAssertEqualObjects(@"", payload[@"exceptions"][0][@"message"]);
-    XCTAssertEqualObjects(event.errorClass,
+    XCTAssertEqualObjects(event.errors[0].errorClass,
                           payload[@"exceptions"][0][@"errorClass"]);
-    XCTAssertEqualObjects(event.errorMessage,
+    XCTAssertEqualObjects(event.errors[0].errorMessage,
                           payload[@"exceptions"][0][@"message"]);
 }
 
@@ -132,9 +132,9 @@
     XCTAssertEqualObjects(@"Exception",
                           payload[@"exceptions"][0][@"errorClass"]);
     XCTAssertEqualObjects(@"", payload[@"exceptions"][0][@"message"]);
-    XCTAssertEqualObjects(event.errorClass,
+    XCTAssertEqualObjects(event.errors[0].errorClass,
                           payload[@"exceptions"][0][@"errorClass"]);
-    XCTAssertEqualObjects(event.errorMessage,
+    XCTAssertEqualObjects(event.errors[0].errorMessage,
                           payload[@"exceptions"][0][@"message"]);
 }
 
@@ -157,9 +157,9 @@
     XCTAssertEqualObjects(@"fatal error",
                           payload[@"exceptions"][0][@"errorClass"]);
     XCTAssertEqualObjects(@"", payload[@"exceptions"][0][@"message"]);
-    XCTAssertEqualObjects(event.errorClass,
+    XCTAssertEqualObjects(event.errors[0].errorClass,
                           payload[@"exceptions"][0][@"errorClass"]);
-    XCTAssertEqualObjects(event.errorMessage,
+    XCTAssertEqualObjects(event.errors[0].errorMessage,
                           payload[@"exceptions"][0][@"message"]);
 }
 
@@ -182,9 +182,9 @@
     XCTAssertEqualObjects(@"assertion failed",
                           payload[@"exceptions"][0][@"errorClass"]);
     XCTAssertEqualObjects(@"", payload[@"exceptions"][0][@"message"]);
-    XCTAssertEqualObjects(event.errorClass,
+    XCTAssertEqualObjects(event.errors[0].errorClass,
                           payload[@"exceptions"][0][@"errorClass"]);
-    XCTAssertEqualObjects(event.errorMessage,
+    XCTAssertEqualObjects(event.errors[0].errorMessage,
                           payload[@"exceptions"][0][@"message"]);
 }
 
@@ -218,9 +218,9 @@
                               payload[@"exceptions"][0][@"errorClass"]);
         XCTAssertEqualObjects(@"Something went wrong",
                               payload[@"exceptions"][0][@"message"]);
-        XCTAssertEqualObjects(event.errorClass,
+        XCTAssertEqualObjects(event.errors[0].errorClass,
                               payload[@"exceptions"][0][@"errorClass"]);
-        XCTAssertEqualObjects(event.errorMessage,
+        XCTAssertEqualObjects(event.errors[0].errorMessage,
                               payload[@"exceptions"][0][@"message"]);
     }
 }
@@ -249,9 +249,9 @@
     XCTAssertEqualObjects(@"fatal error",
                           payload[@"exceptions"][0][@"errorClass"]);
     XCTAssertEqualObjects(@"", payload[@"exceptions"][0][@"message"]);
-    XCTAssertEqualObjects(event.errorClass,
+    XCTAssertEqualObjects(event.errors[0].errorClass,
                           payload[@"exceptions"][0][@"errorClass"]);
-    XCTAssertEqualObjects(event.errorMessage,
+    XCTAssertEqualObjects(event.errors[0].errorMessage,
                           payload[@"exceptions"][0][@"message"]);
 }
 
@@ -279,9 +279,9 @@
     XCTAssertEqualObjects(@"fatal error",
                           payload[@"exceptions"][0][@"errorClass"]);
     XCTAssertEqualObjects(@"", payload[@"exceptions"][0][@"message"]);
-    XCTAssertEqualObjects(event.errorClass,
+    XCTAssertEqualObjects(event.errors[0].errorClass,
                           payload[@"exceptions"][0][@"errorClass"]);
-    XCTAssertEqualObjects(event.errorMessage,
+    XCTAssertEqualObjects(event.errors[0].errorMessage,
                           payload[@"exceptions"][0][@"message"]);
 }
 
@@ -315,9 +315,9 @@
                           payload[@"exceptions"][0][@"errorClass"]);
     XCTAssertEqualObjects(@"A message from beyond | Wo0o0o",
                           payload[@"exceptions"][0][@"message"]);
-    XCTAssertEqualObjects(event.errorClass,
+    XCTAssertEqualObjects(event.errors[0].errorClass,
                           payload[@"exceptions"][0][@"errorClass"]);
-    XCTAssertEqualObjects(event.errorMessage,
+    XCTAssertEqualObjects(event.errors[0].errorMessage,
                           payload[@"exceptions"][0][@"message"]);
 }
 
@@ -345,9 +345,9 @@
     XCTAssertEqualObjects(@"Exception",
                           payload[@"exceptions"][0][@"errorClass"]);
     XCTAssertEqualObjects(@"", payload[@"exceptions"][0][@"message"]);
-    XCTAssertEqualObjects(event.errorClass,
+    XCTAssertEqualObjects(event.errors[0].errorClass,
                           payload[@"exceptions"][0][@"errorClass"]);
-    XCTAssertEqualObjects(event.errorMessage,
+    XCTAssertEqualObjects(event.errors[0].errorMessage,
                           payload[@"exceptions"][0][@"message"]);
 }